+++ CHANGELOG.md
@@ -9,11 +9,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Here we write upgrading notes for brands. It's a team effort to make them as straightforward as possible.
 
-### Fixed 
-- Update createdAt when there is existing URL
-- Create entry when URL does not exist 
+### Fixed
+- Use IReqBody instead of ICreateReqBody
+
+### Deprecated 
+- Deprecate ICreateReqBody
+
+## [0.1.3-beta] - 2022-06-20
+
+### Fixed
+- [a051f22](https://github.com/KarthikUdyawar/url-shortener/commit/a051f22) - Fixed Create short URL API
 
 ## [0.1.2-beta] - 2022-06-20
+
+### Fixed
 - [1735da9](https://github.com/KarthikUdyawar/url-shortener/commit/1735da9) - Fixed RedirectShortUrl API
 
 ---